+++ wtframework/wtf/constants.py
@@ -1,5 +1,5 @@
 ##########################################################################
-#This file is part of WTFramework. 
+# This file is part of WTFramework. 
 #
 #    WTFramework is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -14,4 +14,16 @@
 #    You should have received a copy of the GNU General Public License
 #    along with WTFramework.  If not, see <http://www.gnu.org/licenses/>.
 ##########################################################################
-__VERSION__ = '0.4.4'
+
+###
+# Constants that are used by both WTFramework and by the utility script
+# files.  We're putting the constants here to avoid adding an unnecessary
+# dependency on WTFramework classes inside our utility scripts.
+###
+
+# Constants specifying the config structure.
+WTF_CONFIG_LOCATION = 'configs/'
+WTF_DEFAULT_CONFIG_FILE = 'default'
+WTF_CONFIG_EXT = '.yaml'
+WTF_ENV_VARS = "WTF_ENV"
+