+++ wtframework/wtf/utils/project_utils.py
@@ -1,5 +1,5 @@
 ##########################################################################
-#This file is part of WTFramework. 
+# This file is part of WTFramework. 
 #
 #    WTFramework is free software: you can redistribute it and/or modify
 #    it under the terms of the GNU General Public License as published by
@@ -18,6 +18,8 @@
 import os
 import re
 
+from six import u
+
 
 class ProjectUtils(object):
     """
@@ -25,6 +27,9 @@ class ProjectUtils(object):
     
     """
 
+    WTF_HOME_CONFIG_ENV_VAR = "WTF_HOME"
+
+    __WTF_ROOT_FOLDER_FILE = ".wtf_root_folder"
 
     __root_folder__ = None
 
@@ -42,8 +47,17 @@ class ProjectUtils(object):
         if(cls.__root_folder__ != None):
             return cls.__root_folder__
 
+        # Check for enviornment variable override.
+        try:
+            cls.__root_folder__ = os.environ[cls.WTF_HOME_CONFIG_ENV_VAR]
+        except KeyError:
+            # Means WTF_HOME override isn't specified.
+            pass
+
+        # Search starting from the current working directory and traverse up parent directories for the 
+        # hidden file denoting the project root folder.
         path = os.getcwd()
-        seperator_matches = re.finditer("/|\\\\", path)
+        seperator_matches = re.finditer(u("/|\\\\"), path)
 
         paths_to_search = [path]
         for match in seperator_matches:
@@ -51,10 +65,11 @@ class ProjectUtils(object):
             paths_to_search.insert(0, p)
 
         for path in paths_to_search:
-            if os.path.isfile("{0}/.wtf_root_folder".format(path)):
-                cls.__root_folder__ = path + "/"
+            target_path = os.path.join(path, cls.__WTF_ROOT_FOLDER_FILE)
+            if os.path.isfile(target_path):
+                cls.__root_folder__ = path
                 return cls.__root_folder__
 
-        raise RuntimeError("Missing root project folder locator file '.wtf_root_folder'." \
-                           + "Check to make sure this file is located in your project directory.")
+        raise RuntimeError(u("Missing root project folder locator file '.wtf_root_folder'.") \
+                           + u("Check to make sure this file is located in your project directory."))